anaconda.conf: increase size of eltorito boot image
It calculates the size of efi_img directory with du option
'--apparent-size' to create the eltorito boot image with command mkdosfs.
It is an approximate value which is a little small with the default 512
blocks extra space. The UEFI fireware can't read the boot file in dir
/EFI of the boot image. Increase the BOOTIMG_EXTRA_SPACE to 4096 to make
it work.

Issue: LIN1022-3601

(LOCAL REV: NOT UPSTREAM) -- Send to upstream later

Signed-off-by: Kai Kang <kai.kang@windriver.com>